requirements.txt
agno==2.6.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
docstring_parser==0.18.0
gitdb==4.0.12
GitPython==3.1.48
greenlet==3.5.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.2
playwright==1.58.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
rich==15.0.0
shellingham==1.5.4
smmap==5.0.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
